# Conflicts:
#	bank-service/pom.xml
#	bank-service/src/main/java/com/example/bankservice/client/EmailServiceClient.java
#	bank-service/src/main/java/com/example/bankservice/client/UserServiceClient.java
#	bank-service/src/main/java/com/example/bankservice/domains/dto/CheckEnoughBalanceDto.java
#	bank-service/src/main/java/com/example/bankservice/domains/dto/CreditTransactionDto.java
#	bank-service/src/main/java/com/example/bankservice/domains/dto/TransactionActivationDto.java
#	bank-service/src/main/java/com/example/bankservice/domains/dto/TransactionConfirmDto.java
#	bank-service/src/main/java/com/example/bankservice/domains/dto/TransactionDto.java
#	bank-service/src/main/java/com/example/bankservice/domains/mappers/CreditTransactionMapper.java
#	bank-service/src/main/java/com/example/bankservice/domains/mappers/TransactionMapper.java
#	bank-service/src/main/java/com/example/bankservice/domains/model/CreditTransaction.java
#	bank-service/src/main/java/com/example/bankservice/domains/model/Transaction.java
#	bank-service/src/main/java/com/example/bankservice/domains/model/enums/TransactionState.java
#	bank-service/src/main/java/com/example/bankservice/repositories/CreditTransactionRepository.java
#	bank-service/src/main/java/com/example/bankservice/repositories/TransactionRepository.java
#	bank-service/src/main/java/com/example/bankservice/services/CreditTransactionService.java
#	bank-service/src/main/java/com/example/bankservice/services/TransactionService.java
#	bank-service/src/main/resources/application.properties
#	email-service/src/main/java/com/example/emailservice/controller/TransactionController.java
#	email-service/src/main/java/com/example/emailservice/dto/ConfirmTransactionDTO.java
#	email-service/src/main/java/com/example/emailservice/dto/TransactionActivationDTO.java
#	email-service/src/main/java/com/example/emailservice/model/TransactionActivation.java
#	email-service/src/main/java/com/example/emailservice/repository/TransactionActivationRepository.java
#	email-service/src/main/java/com/example/emailservice/service/impl/TransactionService.java
#	email-service/src/test/java/com/example/emailservice/service/impl/TransactionServiceTest.java
#	pom.xml
#	user-service/src/main/java/rs/edu/raf/userservice/bootstrap/BootstrapData.java
#	user-service/src/main/java/rs/edu/raf/userservice/configuration/SpringSecurityConfig.java
#	user-service/src/main/java/rs/edu/raf/userservice/controllers/AccountController.java
#	user-service/src/main/java/rs/edu/raf/userservice/controllers/ContactController.java
#	user-service/src/main/java/rs/edu/raf/userservice/controllers/CreditController.java
#	user-service/src/main/java/rs/edu/raf/userservice/controllers/CreditRequestController.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/dto/account/CheckEnoughBalanceDto.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/dto/contact/ContactCreateDto.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/dto/contact/ContactDto.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/dto/credit/CreateCreditDto.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/dto/credit/CreditTransactionDto.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/dto/creditrequest/CreditRequestCreateDto.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/dto/creditrequest/CreditRequestDto.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/mappers/ContactMapper.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/mappers/CreditRequestMapper.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/model/Account.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/model/Credit.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/model/CreditRequest.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/model/User.java
#	user-service/src/main/java/rs/edu/raf/userservice/domains/model/enums/CreditRequestStatus.java
#	user-service/src/main/java/rs/edu/raf/userservice/repositories/ContactRepository.java
#	user-service/src/main/java/rs/edu/raf/userservice/services/AccountService.java
#	user-service/src/main/java/rs/edu/raf/userservice/services/ContactService.java
#	user-service/src/main/java/rs/edu/raf/userservice/services/CreditRequestService.java
#	user-service/src/main/java/rs/edu/raf/userservice/services/CreditService.java
#	user-service/src/main/java/rs/edu/raf/userservice/utils/BankServiceClient.java
#	user-service/src/main/resources/application.properties

+++ user-service/src/main/java/rs/edu/raf/userservice/domains/dto/credit/CreditTransactionDto.java
@@ -4,14 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class CreditTransactionDto {
-
     private String accountFrom;
     private Long creditId;
-    private Double amount;
+    private BigDecimal amount;
     private String currencyMark;
     private Long date;
 }
